+++ common/inc/fft_prep_cpx.h
@@ -11,7 +11,7 @@
  *  \return An array storing all of this numbers, must be freed by the user
     @param grade [in] - the grade for which it is computed
 */
-double* every_rou(int);
+long double* every_rou(int);
 
 /**
  *  \brief Compute the first ROU different from (1, 0)
@@ -19,7 +19,7 @@ double* every_rou(int);
     @param grade [in] - the grade for which it is computed
     @param dst [out] - destination at which to save the computed root
 */
-void first_rou(int, double*);
+void first_rou(int, long double*);
 
 
 #endif//_FFT_PREP_CPX_H_
